task 2, fixed attachments and fixed tests
